update delete album et image
ajout de securite de deletealbum et de deleteimage

+++ DOMAINLOGIC/deleteimage.dom.php
@@ -1,8 +1,22 @@
-<?php  
-    include "../CLASSES/IMAGE/image.php";
+<?php
+    
+    include "../CLASSES/ALBUM/album.php";
     include "../CLASSES/COMMENTAIRES/commentaire.php";
+
+    session_start();
+
     $image = new Image();
     $image->load_image($_GET["id"]);
+    $album = new Album();
+    $album->load_album($image->get_authorID());
+
+    if($_SESSION["userID"] != $album->get_authorID()){
+        header("Location: error.php?ErrorMSG=Not%20the%20author!");
+        die();
+    }
+
+    
+    
     $image->delete_image();
     
     $file_pointer = "../".$image->get_imageUrl();  
@@ -14,7 +28,7 @@
     }
 
     $albumID = $image->get_albumID();
-    
+
     header("Location: ../album.php?id=$albumID");
     die();
